+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Update to 1.19.2
+- Updated config lib