+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- Update to 1.19.4
+- Updated config lib